+++ themes/ApplicationTheme_2/assets/js/login.js
@@ -1,0 +1,8 @@
+const $ = require( 'jquery' );
+global.$ = $;
+window.$ = $;
+
+require( '@popperjs/core/dist/umd/popper.js' );
+require( 'bootstrap' );
+
+require( '../vendor/application-login-page/js/main.js' );