+++ webpack.config.js
@@ -69,83 +69,20 @@ adminPanelConfig.name = 'adminPanel';
 //=================================================================================================
 
 /**
- *  Application Default Theme
+ *  Application Theme 1
  */
 Encore.reset();
-Encore
-    .setOutputPath( 'public/__application_slug__/build/default/' )
-    .setPublicPath( '/build/default' )
-   	
-    .autoProvidejQuery()
-    .enableSassLoader(function(sassOptions) {}, {
-        resolveUrlLoader: true
-    })
-    .configureFilenames({
-        js: '[name].js?[contenthash]',
-        css: '[name].css?[contenthash]',
-        assets: '[name].[ext]?[hash:8]'
-    })
-    .enableSingleRuntimeChunk()
-    .enableVersioning(Encore.isProduction())
-    .enableSourceMaps( !Encore.isProduction() )
-    
-    .copyFiles({
-         from: './assets/__application_slug__/images',
-         to: 'images/[path][name].[ext]',
-     })
-    
-    // Add Entries
-    .addStyleEntry( 'css/login', './assets/__application_slug__/css/login.css' )
-    .addEntry( 'js/login', './assets/__application_slug__/js/login.js' )
-    
-    .addEntry( 'js/home', './assets/__application_slug__/js/pages/home.js' )
-;
-
-const applicationConfig = Encore.getWebpackConfig();
-applicationConfig.name = '__application_slug__';
+const applicationTheme1Config   = require('./themes/ApplicationTheme_1/webpack.config');
 
 //=================================================================================================
 
 /**
- *  Test Theme
+ *  Application Theme 2
  */
- Encore.reset();
-Encore
-    .setOutputPath( 'public/__application_slug__/build/test-theme/' )
-    .setPublicPath( '/build/test-theme' )
-    
-    .autoProvidejQuery()
-    .enableSassLoader(function(sassOptions) {}, {
-        resolveUrlLoader: true
-    })
-    .configureFilenames({
-        js: '[name].js?[contenthash]',
-        css: '[name].css?[contenthash]',
-        assets: '[name].[ext]?[hash:8]'
-    })
-    .enableSingleRuntimeChunk()
-    .enableVersioning(Encore.isProduction())
-    .enableSourceMaps( !Encore.isProduction() )
-    
-    .copyFiles({
-         from: './assets/test-theme/images',
-         to: 'images/[path][name].[ext]',
-     })
-    
-    // Add Entries
-    .addStyleEntry( 'css/login', './assets/test-theme/css/login.scss' )
-    .addEntry( 'js/login', './assets/test-theme/js/pages/login.js' )
-    
-    .addStyleEntry( 'css/app', './assets/test-theme/css/main.scss' )
-    .addEntry( 'js/app', './assets/test-theme/js/app.js' )
-    
-    .addEntry( 'js/home', './assets/test-theme/js/pages/home.js' )
-;
-
-const testThemeConfig = Encore.getWebpackConfig();
-testThemeConfig.name = 'test-theme';
+Encore.reset();
+const applicationTheme2Config   = require('./themes/ApplicationTheme_2/webpack.config');
 
 //=================================================================================================
 
 
-module.exports = [adminPanelConfig, applicationConfig, testThemeConfig];
+module.exports = [adminPanelConfig, applicationTheme1Config, applicationTheme2Config];